Exclude spec files from generated assets
The commit makes changes to the tsconfig.json files in both the 'adapter-next-kv' and 'core' packages. Modified the inclusion pattern to only include '*.ts' files and added an 'exclude' pattern to avoid unnecessary compilation of 'node_modules' and any spec files in 'ts' directory. This was done to improve build performance and skip irrelevant files.