Exclude fixtures from output
The versions of edgesession and @edgesession/adapter-next-kv have been incremented to 0.0.3 to reflect recent changes to the packages. The specified packages are now updated in the package dependencies as well. Also, a change has been added to exclude __fixtures__ from the TypeScript compile process in tsconfig.build.json, reducing unnecessary compile time in the development environment.